+++ tests/eval_files/async911.py
@@ -1,5 +1,4 @@
 # AUTOFIX
-# ASYNCIO_NO_AUTOFIX
 from typing import Any
 
 import pytest
@@ -7,11 +6,11 @@ import trio
 
 _: Any = ""
 
-# ARG --enable=ASYNC910,ASYNC911
+# ARG --enable=ASYNC910,ASYNC911,ASYNC914
 
 
 async def foo() -> Any:
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 def bar(*args) -> Any: ...
@@ -22,19 +21,19 @@ def condition() -> Any: ...
 
 
 async def foo_yield_1():
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield 5
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 async def foo_yield_2():
     yield  # error: 4, "yield", Statement("function definition", lineno-1)
     yield  # error: 4, "yield", Statement("yield", lineno-1)
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 async def foo_yield_3():  # error: 0, "exit", Statement("yield", lineno+2)
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
 
 
@@ -50,15 +49,15 @@ async def foo_yield_return_1():
 
 
 async def foo_yield_return_2():
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
     return  # error: 4, "return", Statement("yield", lineno-1)
 
 
 async def foo_yield_return_3():
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
-    await foo()
+    await trio.lowlevel.checkpoint()
     return
 
 
@@ -104,14 +103,14 @@ async def foo_async_for_4():  # safe
 
 # for
 async def foo_for():  # error: 0, "exit", Statement("yield", lineno+3)
-    await foo()
+    await trio.lowlevel.checkpoint()
     for i in "":
         yield  # error: 8, "yield", Statement("yield", lineno)
 
 
 async def foo_for_1():  # error: 0, "exit", Statement("function definition", lineno) # error: 0, "exit", Statement("yield", lineno+3)
     for _ in "":
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
 
 
@@ -123,13 +122,13 @@ async def foo_while_1():  # error: 0, "exit", Statement("yield", lineno+5)
     while foo():
         ...
     else:
-        await foo()  # will always run
+        await trio.lowlevel.checkpoint()  # will always run
     yield  # safe
 
 
 # simple yield-in-loop case
 async def foo_while_2():  # error: 0, "exit", Statement("yield", lineno+3)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # error: 8, "yield", Statement("yield", lineno)
 
@@ -137,7 +136,7 @@ async def foo_while_2():  # error: 0, "exit", Statement("yield", lineno+3)
 # no checkpoint after yield if else is entered
 async def foo_while_3():  # error: 0, "exit", Statement("yield", lineno+5)
     while foo():
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
     else:
         yield  # error: 8, "yield", Statement("yield", lineno-2) # error: 8, "yield", Statement("function definition", lineno-5)
@@ -145,7 +144,7 @@ async def foo_while_3():  # error: 0, "exit", Statement("yield", lineno+5)
 
 # check that errors are suppressed in visit_While
 async def foo_while_4():  # error: 0, "exit", Statement("yield", lineno+3) # error: 0, "exit", Statement("yield", lineno+5) # error: 0, "exit", Statement("yield", lineno+7)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # error: 8, "yield", Statement("yield", lineno) # error: 8, "yield", Statement("yield", lineno+2) # error: 8, "yield", Statement("yield", lineno+4)
         while foo():
@@ -156,7 +155,7 @@ async def foo_while_4():  # error: 0, "exit", Statement("yield", lineno+3) # err
 
 # check that state management is handled in for loops as well
 async def foo_while_4_for():  # error: 0, "exit", Statement("yield", lineno+3) # error: 0, "exit", Statement("yield", lineno+5) # error: 0, "exit", Statement("yield", lineno+7)
-    await foo()
+    await trio.lowlevel.checkpoint()
     for i in bar():
         yield  # error: 8, "yield", Statement("yield", lineno) # error: 8, "yield", Statement("yield", lineno+2) # error: 8, "yield", Statement("yield", lineno+4)
         for i in bar():
@@ -167,40 +166,40 @@ async def foo_while_4_for():  # error: 0, "exit", Statement("yield", lineno+3) #
 
 # check error suppression is reset
 async def foo_while_5():
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # error: 8, "yield", Statement("yield", lineno)
 
         async def foo_nested_error():  # error: 8, "exit", Statement("yield", lineno+1)
             yield  # error: 12, "yield", Statement("function definition", lineno-1)
 
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 # --- while + continue ---
 # no checkpoint on continue
 async def foo_while_continue_1():  # error: 0, "exit", Statement("yield", lineno+3)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # error: 8, "yield", Statement("yield", lineno)
         if condition():
             continue
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 # multiple continues
 async def foo_while_continue_2():  # error: 0, "exit", Statement("yield", lineno+3)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # error: 8, "yield", Statement("yield", lineno)
         if foo():
             continue
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             continue
         while foo():
             yield  # safe
-            await foo()
+            await trio.lowlevel.checkpoint()
 
 
 # --- while + break ---
@@ -210,28 +209,28 @@ async def foo_while_break_1():  # error: 0, "exit", Statement("yield", lineno+6)
         if condition():
             break
     else:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Statement("function definition", lineno-6)
 
 
 # no checkpoint on break
 async def foo_while_break_2():  # error: 0, "exit", Statement("yield", lineno+3)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield  # safe
         if condition():
             break
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 # guaranteed if else and break
 async def foo_while_break_3():  # error: 0, "exit", Statement("yield", lineno+7)
     while foo():
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break  # if it breaks, have checkpointed
     else:
-        await foo()  # runs if 0-iter
+        await trio.lowlevel.checkpoint()  # runs if 0-iter
     yield  # safe
 
 
@@ -240,38 +239,38 @@ async def foo_while_break_4():  # error: 0, "exit", Statement("yield", lineno+7)
     while foo():
         if condition():
             break
-        await foo()  # might not run
+        await trio.lowlevel.checkpoint()  # might not run
     else:
-        await foo()  # might not run
+        await trio.lowlevel.checkpoint()  # might not run
     yield  # error: 4, "yield", Statement("function definition", lineno-7)
 
 
 # check break is reset on nested
 async def foo_while_break_5():  # error: 0, "exit", Statement("yield", lineno+12)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield
         if condition():
             break
-        await foo()
+        await trio.lowlevel.checkpoint()
         while foo():
             yield  # safe
-            await foo()
+            await trio.lowlevel.checkpoint()
         yield  # safe
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Statement("yield", lineno-9)
 
 
 # check multiple breaks
 async def foo_while_break_6():  # error: 0, "exit", Statement("yield", lineno+11)
-    await foo()
+    await trio.lowlevel.checkpoint()
     while foo():
         yield
         if condition():
             break
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
     yield  # error: 4, "yield", Statement("yield", lineno-8)
@@ -279,7 +278,7 @@ async def foo_while_break_6():  # error: 0, "exit", Statement("yield", lineno+11
 
 async def foo_while_break_7():  # error: 0, "exit", Statement("function definition", lineno)# error: 0, "exit", Statement("yield", lineno+5)
     while foo():
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
         yield
@@ -288,13 +287,13 @@ async def foo_while_break_7():  # error: 0, "exit", Statement("function definiti
 
 async def foo_while_endless_1():
     while True:
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
 
 
 async def foo_while_endless_2():  # error: 0, "exit", Statement("function definition", lineno)# error: 0, "exit", Statement("yield", lineno+3)
     while foo():
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
 
 
@@ -302,15 +301,15 @@ async def foo_while_endless_3():
     while True:
         ...
     yield  # type: ignore[unreachable]
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 async def foo_while_endless_4():
-    await foo()
+    await trio.lowlevel.checkpoint()
     while True:
         yield
         while True:
-            await foo()
+            await trio.lowlevel.checkpoint()
             yield
 
 
@@ -325,12 +324,12 @@ async def foo_try_1():  # error: 0, "exit", Statement("function definition", lin
 # no checkpoint after yield in ValueError
 async def foo_try_2():  # error: 0, "exit", Statement("yield", lineno+5)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except ValueError:
         # try might not have checkpointed
         yield  # error: 8, "yield", Statement("function definition", lineno-5)
     except:
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         pass
 
@@ -339,7 +338,7 @@ async def foo_try_3():  # error: 0, "exit", Statement("yield", lineno+6)
     try:
         ...
     except:
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield  # error: 8, "yield", Statement("function definition", lineno-6)
 
@@ -350,38 +349,38 @@ async def foo_try_4():  # safe
     except:
         yield  # error: 8, "yield", Statement("function definition", lineno-4)
     finally:
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_try_5():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         # try might crash before checkpoint
         yield  # error: 8, "yield", Statement("function definition", lineno-5)
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_try_6():  # error: 0, "exit", Statement("yield", lineno+5)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except ValueError:
         pass
     yield  # error: 4, "yield", Statement("function definition", lineno-5)
 
 
 async def foo_try_7():  # error: 0, "exit", Statement("yield", lineno+17)
-    await foo()
+    await trio.lowlevel.checkpoint()
     try:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
     except ValueError:
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
     except SyntaxError:
         yield  # error: 8, "yield", Statement("yield", lineno-7)
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         pass
     # If the try raises an exception without checkpointing, and it's not caught
@@ -391,26 +390,36 @@ async def foo_try_7():  # error: 0, "exit", Statement("yield", lineno+17)
 
 
 ## safe only if (try or else) and all except bodies either await or raise
-## if foo() raises a ValueError it's not checkpointed, and may or may not yield
-async def foo_try_8():  # error: 0, "exit", Statement("function definition", lineno) # error: 0, "exit", Statement("yield", lineno+3)
+## if the await raises an exception it's not checkpointed, and may or may not yield
+async def foo_try_8a():  # error: 0, "exit", Statement("function definition", lineno) # error: 0, "exit", Statement("yield", lineno+3)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
-        await foo()
     except ValueError:
         ...
     except:
         raise
     else:
-        await foo()
+        await trio.lowlevel.checkpoint()
+
+
+async def foo_try_8b():  # error: 0, "exit", Statement("function definition", lineno) # error: 0, "exit", Statement("yield", lineno+3)
+    try:
+        await trio.lowlevel.checkpoint()
+        yield
+        await trio.lowlevel.checkpoint()
+    except ValueError:
+        ...
+    except:
+        raise
 
 
 # no checkpoint after yield in else
 async def foo_try_9():  # error: 0, "exit", Statement("yield", lineno+6)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except:
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield
 
@@ -418,56 +427,56 @@ async def foo_try_9():  # error: 0, "exit", Statement("yield", lineno+6)
 # bare except means we'll jump to finally after full execution of either try or the except
 async def foo_try_10():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except:
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_try_10_BaseException():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except BaseException:
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 # not fully covering excepts
 async def foo_try_10_exception():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except ValueError:
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         yield  # error: 8, "yield", Statement("function definition", lineno-6)
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_try_10_no_except():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     finally:
         # try might crash before checkpoint
         yield  # error: 8, "yield", Statement("function definition", lineno-5)
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 # if
 async def foo_if_1():
     if condition():
         yield  # error: 8, "yield", Statement("function definition", lineno-2)
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield  # error: 8, "yield", Statement("function definition", lineno-5)
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_if_2():  # error: 0, "exit", Statement("yield", lineno+6)
-    await foo()
+    await trio.lowlevel.checkpoint()
     if condition():
         ...
     else:
@@ -476,7 +485,7 @@ async def foo_if_2():  # error: 0, "exit", Statement("yield", lineno+6)
 
 
 async def foo_if_3():  # error: 0, "exit", Statement("yield", lineno+6)
-    await foo()
+    await trio.lowlevel.checkpoint()
     if condition():
         yield
     else:
@@ -485,20 +494,20 @@ async def foo_if_3():  # error: 0, "exit", Statement("yield", lineno+6)
 
 
 async def foo_if_4():  # error: 0, "exit", Statement("yield", lineno+7)
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
     if condition():
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         ...
     yield  # error: 4, "yield", Statement("yield", lineno-5)
 
 
 async def foo_if_5():  # error: 0, "exit", Statement("yield", lineno+8)
-    await foo()
+    await trio.lowlevel.checkpoint()
     if condition():
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield
         ...
@@ -506,30 +515,30 @@ async def foo_if_5():  # error: 0, "exit", Statement("yield", lineno+8)
 
 
 async def foo_if_6():  # error: 0, "exit", Statement("yield", lineno+8)
-    await foo()
+    await trio.lowlevel.checkpoint()
     if condition():
         yield
     else:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
         ...
     yield  # error: 4, "yield", Statement("yield", lineno-5)
 
 
 async def foo_if_7():  # error: 0, "exit", Statement("function definition", lineno)
     if condition():
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_if_8():  # error: 0, "exit", Statement("function definition", lineno)
     if condition():
         ...
     else:
-        await foo()
+        await trio.lowlevel.checkpoint()
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 # IfExp
@@ -585,7 +594,7 @@ def foo_sync_7():
 
 # nested function definition
 async def foo_func_1():
-    await foo()
+    await trio.lowlevel.checkpoint()
 
     async def foo_func_2():  # error: 4, "exit", Statement("yield", lineno+1)
         yield  # error: 8, "yield", Statement("function definition", lineno-1)
@@ -595,22 +604,22 @@ async def foo_func_1():
 # so we need to disable black
 # fmt: off
 async def foo_func_3():  # error: 0, "exit", Statement("yield", lineno+2)
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
 
     async def foo_func_4():
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_func_5():  # error: 0, "exit", Statement("yield", lineno+2)
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
 
     def foo_func_6():  # safe
         yield
 
         async def foo_func_7():
-            await foo()
+            await trio.lowlevel.checkpoint()
             ...
 # fmt: on
 
@@ -624,161 +633,161 @@ async def foo_boolops_1():  # error: 0, "exit", Stmt("yield", line+1)
 async def foo_loop_static():
     # break/else behaviour on guaranteed body execution
     for _ in [1, 2, 3]:
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
 
     for _ in [1, 2, 3]:
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
     else:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     # continue
     for _ in [1, 2, 3]:
         if condition():
             continue
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-7)
 
     # continue/else
     for _ in [1, 2, 3]:
         if condition():
             continue
-        await foo()
+        await trio.lowlevel.checkpoint()
     else:
         yield  # error: 8, "yield", Stmt("yield", line-8)
-    await foo()
+    await trio.lowlevel.checkpoint()
     yield
 
     for _ in [1, 2, 3]:
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
     else:
         yield
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     # test different containers
     for _ in (1, 2, 3):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in (foo(), foo()):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in ((),):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in "hello":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in b"hello":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in r"hello":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in {1, 2, 3}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in ():
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in {1: 2, 3: 4}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in "   ".strip():
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in range(0):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in (*range(0),):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in (*(1, 2),):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in {**{}}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in {**{}, **{}}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in {**{1: 2}}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in (*range(0), *[1, 2, 3]):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in {**{}, **{1: 2}}:  # type: ignore[arg-type]
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     x: Any = ...
     for _ in (*x, *[1, 2, 3]):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in {**x, **{1: 2}}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for _ in {}:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in "":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in """""":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in [[], []][0]:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for _ in [[], []].__getitem__(0):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     # not handled
     for _ in list((1, 2)):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-5)
 
     for _ in list():
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     # while
     while True:
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
     yield
@@ -786,102 +795,102 @@ async def foo_loop_static():
     while True:
         if condition():
             break
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-6)
 
     while True:
         if condition():
             continue
-        await foo()
+        await trio.lowlevel.checkpoint()
         if condition():
             break
     yield
 
     while False:
-        await foo()  # type: ignore[unreachable]
+        await trio.lowlevel.checkpoint()  # type: ignore[unreachable]
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     while "hello":
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     # false positive on containers
     while [1, 2]:
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-5)
 
     # will get caught by any number of linters, but trio911 will also complain
     for _ in 5:  # type: ignore
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-5)
 
     # range with constant arguments also handled, see more extensive tests in 910
     for i in range(5):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(0x23):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(0b01):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(1 + 1):  # not handled
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for i in range(None):  # type: ignore
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     for i in range(+3):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(-3.5):  # type: ignore
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     # duplicated from 910 to have all range tests in one place
     for i in range(5, 10):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(10, 5, -1):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     # ~0 == -1
     for i in range(10, 5, ~0):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     # length > sys.maxsize
     for i in range(27670116110564327421):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield
 
     for i in range(10, 5):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     # binary operations are not handled
     for i in range(3 - 2):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-5)
 
     for i in range(10**3):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-4)
 
     # nor nested unary operations
     for i in range(--3):
-        await foo()
+        await trio.lowlevel.checkpoint()
     yield  # error: 4, "yield", Stmt("yield", line-5)
 
-    await foo()
+    await trio.lowlevel.checkpoint()
 
 
 # don't warn on pytest.fixture