+++ tests/eval_files/noqa_testing.py
@@ -1,7 +1,6 @@
 # TODO: When was this file added? Why?
 
 # AUTOFIX
-# ASYNCIO_NO_AUTOFIX
 # ARG --enable=ASYNC911
 import trio